move infinite scroll table to web-common table
