+++ web-admin/src/features/organizations/users/OrgUsersTable.svelte
@@ -15,7 +15,7 @@
     InfiniteQueryObserverResult,
   } from "@tanstack/svelte-query";
   import { ExternalLinkIcon } from "lucide-svelte";
-  import InfiniteScrollTable from "@rilldata/web-admin/components/InfiniteScrollTable.svelte";
+  import InfiniteScrollTable from "@rilldata/web-common/components/table/InfiniteScrollTable.svelte";
 
   interface OrgUser extends V1OrganizationMemberUser, V1OrganizationInvite {
     invitedBy?: string;
